added functions to login and display task list

+++ www/js/index.js
@@ -1,34 +1,49 @@
+
 $(document).ready(function () {
     $("#addButton").click(function () {
-        var meter = $("#meterId").val()
-        $.ajax({
-            url: 'http://localhost:3000',
-            type: 'POST',
-            data: {
-                'meterId': $("#meterId").val(),
-                'model': $("#model").val(),
-                'custId': $("#custid").val()
-            },
-            success: function (response) {
-                $("#content").text(response.result);
-            }
-        })
-
+        var merg_url = localStorage.getItem("merg_url");
+        if (merg_url === null)
+            alert('please configure end point')
+        else {
+            var meter = $("#meterId").val()
+            $.ajax({
+                url: merg_url,
+                type: 'POST',
+                data: {
+                    'meterId': $("#meterId").val(),
+                    'model': $("#model").val(),
+                    'custId': $("#custid").val()
+                },
+                success: function (response) {
+                    $("#content").text(response.result);
+                }
+            })
+        }
     })
     $("#delButton").click(function () {
-        var meter = $("#meterId").val()
-        $.ajax({
-            url: 'http://localhost:3000',
-            type: 'DELETE',
-            data: { 'meterId': meter },
-            success: function (response) {
-                $("#content").text(response.result);
-            }
-        })
-
+        var merg_url = localStorage.getItem("merg_url");
+        if (merg_url === null)
+            alert('please configure end point')
+        else {
+            var meter = $("#meterId").val()
+            $.ajax({
+                url: merg_url,
+                type: 'DELETE',
+                data: { 'meterId': meter },
+                success: function (response) {
+                    $("#content").text(response.result);
+                }
+            })
+        }
     })
+
     $("#cfgButton").click(function () {
-        var hub = $("#hub").val()
-        $("#content").text('not yet implemented');
+        localStorage.setItem("merg_url", $("#merg_url").val());
+        $("#content").text('enpoint configured');
+    })
+
+    $("#loginButton").click(function () {
+        localStorage.setItem("agentId", $("#agId").val());
+        $("#content").text("agent " + localStorage.getItem("agentId") + " succesfully logged in");
     })
 })
